requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-cohere
faiss-cpu
faiss-gpu
PyPDF2
python-dotenv
torch
transformers
rank_bm25
sentence-transformers
langchain-core
python-docx
ollama
